exec refactored until exec_chain.c
