+++ include/minishell.h
@@ -1,7 +1,6 @@
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
-// # include "libft.h"
 # include "arraydef.h"
 # include "cmddef.h"
 
@@ -12,5 +11,20 @@ int		read_input(char **str);
 
 void	signals(void);
 void	ctrlc(int sig);
+
+int		exec_chain(t_chain *chain);
+void	exec_sequence(t_array *sequence);
+void	exec_external(t_command *cmd);
+int		exec_builtin(t_command *command);
+
+void	run_child(t_command *cmd, int input, int ports[2], int last);
+void	run_parent(t_command *cmd, int *fd, int ports[2]);
+int		run_builtin_in_main(t_command *cmd);
+
+int		super_duper(int fd_src, int fd_dst);
+int		redirect(t_command *cmd, int input, int ports[2], int last);
+
+int		cmd_is_external(t_command *cmd);
+int		is_local_script(char *cmd);
 
 #endif
